requirements.txt
Django>=3.0.3,<3.1.0
djangorestframework>=3.11.0,<3.12.0

flake8>=3.7.9,<3.8.0

mysqlclient>=1.4.6,<1.5.0
django-mysql>=3.3.0,<3.4.0
celery>=4.4.0,<4.5.0
pysrt>=1.1.2,<1.2.0